fix: statically build correct nav for each section
