+++ apps/docs/components/reference/RefSectionHandler.tsx
@@ -10,6 +10,7 @@ import ApiOperationSection from './ApiOperationSection'
 import CliCommandSection from './CLICommandSection'
 import OldVersionAlert from './OldVersionAlert'
 import { IAPISpec, ICommonSection, IRefStaticDoc, ISpec, TypeSpec } from './Reference.types'
+import { MainSkeleton } from '~/layouts/MainSkeleton'
 
 interface RefSectionHandlerProps {
   sections: ICommonSection[]
@@ -18,6 +19,7 @@ interface RefSectionHandlerProps {
   pageProps: { docs: IRefStaticDoc[] }
   type: 'client-lib' | 'cli' | 'api'
   isOldVersion?: boolean
+  menuId: string
 }
 
 const RefSectionHandler = (props: RefSectionHandlerProps) => {
@@ -75,53 +77,55 @@ const RefSectionHandler = (props: RefSectionHandlerProps) => {
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <link rel="canonical" href={`https://supabase.com${router.basePath}${path}`} />
       </Head>
-      {props.isOldVersion && <OldVersionAlert sections={props.sections} />}
-      <RefSubLayout>
-        {props.sections.map((section, i) => {
-          const sectionType = section.type
-          switch (sectionType) {
-            case 'markdown':
-              const markdownData = props.pageProps.docs.find((doc) => doc.id === section.id)
+      <MainSkeleton menuId={props.menuId}>
+        {props.isOldVersion && <OldVersionAlert sections={props.sections} />}
+        <RefSubLayout>
+          {props.sections.map((section, i) => {
+            const sectionType = section.type
+            switch (sectionType) {
+              case 'markdown':
+                const markdownData = props.pageProps.docs.find((doc) => doc.id === section.id)
 
-              return (
-                <RefEducationSection
-                  key={section.id + i}
-                  item={section}
-                  markdownContent={markdownData}
-                />
-              )
-            case 'function':
-              return (
-                <RefFunctionSection
-                  key={section.id + i}
-                  funcData={section}
-                  commonFuncData={section}
-                  spec={props.spec}
-                  typeSpec={props.typeSpec}
-                />
-              )
-            case 'cli-command':
-              return (
-                <CliCommandSection
-                  key={section.id + i}
-                  funcData={section}
-                  commonFuncData={section}
-                />
-              )
-            case 'operation':
-              return (
-                <ApiOperationSection
-                  key={section.id + i}
-                  funcData={section}
-                  commonFuncData={section}
-                  spec={props.spec}
-                />
-              )
-            default:
-              throw new Error(`Unknown common section type '${sectionType}'`)
-          }
-        })}
-      </RefSubLayout>
+                return (
+                  <RefEducationSection
+                    key={section.id + i}
+                    item={section}
+                    markdownContent={markdownData}
+                  />
+                )
+              case 'function':
+                return (
+                  <RefFunctionSection
+                    key={section.id + i}
+                    funcData={section}
+                    commonFuncData={section}
+                    spec={props.spec}
+                    typeSpec={props.typeSpec}
+                  />
+                )
+              case 'cli-command':
+                return (
+                  <CliCommandSection
+                    key={section.id + i}
+                    funcData={section}
+                    commonFuncData={section}
+                  />
+                )
+              case 'operation':
+                return (
+                  <ApiOperationSection
+                    key={section.id + i}
+                    funcData={section}
+                    commonFuncData={section}
+                    spec={props.spec}
+                  />
+                )
+              default:
+                throw new Error(`Unknown common section type '${sectionType}'`)
+            }
+          })}
+        </RefSubLayout>
+      </MainSkeleton>
     </>
   )
 }
