+++ apps/docs/layouts/guides/index.tsx
@@ -12,6 +12,7 @@ import { FooterHelpCalloutType } from '~/components/FooterHelpCallout'
 import GuidesTableOfContents from '~/components/GuidesTableOfContents'
 import useHash from '~/hooks/useHash'
 import { LayoutMainContent } from '../DefaultLayout'
+import { MainSkeleton } from '../MainSkeleton'
 
 interface Props {
   meta: {
@@ -30,6 +31,7 @@ interface Props {
   toc?: any
   currentPage?: string
   hideToc?: boolean
+  menuId: string
 }
 
 const Layout: FC<Props> = (props) => {
@@ -41,6 +43,8 @@ const Layout: FC<Props> = (props) => {
 
   const { asPath } = useRouter()
   const router = useRouter()
+
+  const menuId = props.menuId
 
   const EDIT_BUTTON_EXCLUDE_LIST = ['/404']
 
@@ -108,77 +112,79 @@ const Layout: FC<Props> = (props) => {
           },
         }}
       />
-      <LayoutMainContent className="pb-0">
-        <div className={['grid grid-cols-12 relative gap-4'].join(' ')}>
-          <div
-            className={[
-              'relative',
-              !props.hideToc ? 'col-span-12 md:col-span-9' : 'col-span-12',
-              'transition-all ease-out',
-              'duration-100',
-            ].join(' ')}
-          >
-            {props.meta?.breadcrumb && (
-              <p className="text-brand tracking-wider mb-3">{props.meta.breadcrumb}</p>
-            )}
-            <article
-              ref={articleRef}
-              className={`${
-                props.meta?.hide_table_of_contents || !hasTableOfContents ? '' : ''
-              } prose max-w-none`}
-            >
-              <h1 className="mb-0">{props.meta.title}</h1>
-              {props.meta?.subtitle && (
-                <h2 className="mt-3 text-xl text-foreground-light">{props.meta.subtitle}</h2>
-              )}
-              <div className="w-full border-b my-8"></div>
-              <MDXProvider components={components}>{props.children}</MDXProvider>
-
-              {EDIT_BUTTON_EXCLUDE_LIST.includes(router.route) ? (
-                <></>
-              ) : (
-                <div className="mt-16 not-prose">
-                  <div>
-                    <a
-                      href={`https://github.com/${
-                        props.editLink ||
-                        `supabase/supabase/edit/master/apps/docs/pages${router.asPath}.mdx`
-                      }
-                    `}
-                      className="text-sm transition flex items-center gap-1 text-scale-1000 hover:text-scale-1200 w-fit"
-                    >
-                      Edit this page on GitHub <IconExternalLink size={14} strokeWidth={1.5} />
-                    </a>
-                  </div>
-                </div>
-              )}
-            </article>
-          </div>
-          {!props.hideToc && hasTableOfContents && !props.meta?.hide_table_of_contents && (
+      <MainSkeleton menuId={menuId}>
+        <LayoutMainContent className="pb-0">
+          <div className={['grid grid-cols-12 relative gap-4'].join(' ')}>
             <div
               className={[
-                'col-span-3',
-                'border-overlay bg-background table-of-contents-height',
-                'thin-scrollbar overflow-y-auto sticky hidden md:block md:col-span-3 px-2',
+                'relative',
+                !props.hideToc ? 'col-span-12 md:col-span-9' : 'col-span-12',
                 'transition-all ease-out',
                 'duration-100',
               ].join(' ')}
             >
-              <div className="border-l">
-                {props.meta?.tocVideo && !!tocVideoPreview && (
-                  <div className="relative mb-6 pl-5">
-                    <ExpandableVideo imgUrl={tocVideoPreview} videoId={props.meta.tocVideo} />
+              {props.meta?.breadcrumb && (
+                <p className="text-brand tracking-wider mb-3">{props.meta.breadcrumb}</p>
+              )}
+              <article
+                ref={articleRef}
+                className={`${
+                  props.meta?.hide_table_of_contents || !hasTableOfContents ? '' : ''
+                } prose max-w-none`}
+              >
+                <h1 className="mb-0">{props.meta.title}</h1>
+                {props.meta?.subtitle && (
+                  <h2 className="mt-3 text-xl text-foreground-light">{props.meta.subtitle}</h2>
+                )}
+                <div className="w-full border-b my-8"></div>
+                <MDXProvider components={components}>{props.children}</MDXProvider>
+
+                {EDIT_BUTTON_EXCLUDE_LIST.includes(router.route) ? (
+                  <></>
+                ) : (
+                  <div className="mt-16 not-prose">
+                    <div>
+                      <a
+                        href={`https://github.com/${
+                          props.editLink ||
+                          `supabase/supabase/edit/master/apps/docs/pages${router.asPath}.mdx`
+                        }
+                    `}
+                        className="text-sm transition flex items-center gap-1 text-scale-1000 hover:text-scale-1200 w-fit"
+                      >
+                        Edit this page on GitHub <IconExternalLink size={14} strokeWidth={1.5} />
+                      </a>
+                    </div>
                   </div>
                 )}
-                <span className="block font-mono text-xs uppercase text-foreground px-5 mb-6">
-                  On this page
-                </span>
-                <GuidesTableOfContents list={tocList} />
-              </div>
+              </article>
             </div>
-          )}
-        </div>
-      </LayoutMainContent>
+            {!props.hideToc && hasTableOfContents && !props.meta?.hide_table_of_contents && (
+              <div
+                className={[
+                  'col-span-3',
+                  'border-overlay bg-background table-of-contents-height',
+                  'thin-scrollbar overflow-y-auto sticky hidden md:block md:col-span-3 px-2',
+                  'transition-all ease-out',
+                  'duration-100',
+                ].join(' ')}
+              >
+                <div className="border-l">
+                  {props.meta?.tocVideo && !!tocVideoPreview && (
+                    <div className="relative mb-6 pl-5">
+                      <ExpandableVideo imgUrl={tocVideoPreview} videoId={props.meta.tocVideo} />
+                    </div>
+                  )}
+                  <span className="block font-mono text-xs uppercase text-foreground px-5 mb-6">
+                    On this page
+                  </span>
+                  <GuidesTableOfContents list={tocList} />
+                </div>
+              </div>
+            )}
+          </div>
+        </LayoutMainContent>
+      </MainSkeleton>
     </>
   )
 }
