+++ apps/docs/pages/guides/api/[[...slug]].tsx
@@ -21,7 +21,7 @@ export default function ApiGuide({
   const { hideToc, ...meta } = frontmatter
 
   return (
-    <Layout meta={meta} hideToc={hideToc} editLink={editLink}>
+    <Layout meta={meta} hideToc={hideToc} editLink={editLink} menuId="api">
       <MDXRemote {...mdxSource} components={components} />
     </Layout>
   )
