+++ apps/docs/pages/guides/cli/[[...slug]].tsx
@@ -21,7 +21,7 @@ export default function CliGuide({
   const { hideToc, ...meta } = frontmatter
 
   return (
-    <Layout meta={meta} hideToc={hideToc} editLink={editLink}>
+    <Layout meta={meta} hideToc={hideToc} editLink={editLink} menuId="supabase_cli">
       <MDXRemote {...mdxSource} components={components} />
     </Layout>
   )
