+++ apps/docs/pages/guides/realtime/[[...slug]].tsx
@@ -21,7 +21,7 @@ export default function RealtimeGuide({
   const { hideToc, ...meta } = frontmatter
 
   return (
-    <Layout meta={meta} hideToc={hideToc} editLink={editLink}>
+    <Layout meta={meta} hideToc={hideToc} editLink={editLink} menuId="realtime">
       <MDXRemote {...mdxSource} components={components} />
     </Layout>
   )
