--- apps/docs/pages/handbook/contributing.mdx
@@ -1,32 +1,0 @@
-import Layout from '~/layouts/DefaultGuideLayout'
-
-export const meta = {
-  id: 'contributing',
-  title: 'Contributing',
-  description: 'Want to help?',
-}
-
-## How to contribute
-
-Supabase has many [open source repos](https://github.com/orgs/supabase/repositories). Some of the more popular ones are:
-
-- [Supabase](https://github.com/supabase/supabase) - Our Dashboard, Websites, and example apps
-- [Supabase-js](https://github.com/supabase/supabase-js) - A client-side library written in TypeScript
-- [Gotrue](https://github.com/supabase/gotrue) - Our auth server, written in Go
-- [Realtime](https://github.com/supabase/realtime) - A server for streaming changes from Postgres, written in Elixir
-- [browse all](https://github.com/orgs/supabase/repositories)
-
-Want to contribute? Why not jump into our GitHub repo and:
-
-- [Sponsor Supabase](https://github.com/sponsors/supabase).
-- [Answer Discussions](https://github.com/supabase/supabase/discussions).
-- Submit an issue to one of our [repos](https://github.com/orgs/supabase/repositories)
-- Report a performance issue or a part of the documentation that you find confusing.
-- Create a pull request in one of our [repos](https://github.com/orgs/supabase/repositories)
-- [Translate our Readme](https://github.com/supabase/supabase/issues/1341).
-- Try our products and give feedback.
-- Spread the word if you like what we are doing.
-
-export const Page = ({ children }) => <Layout meta={meta} children={children} />
-
-export default Page
