--- apps/docs/pages/handbook/introduction.mdx
@@ -1,21 +1,0 @@
-import Layout from '~/layouts/DefaultGuideLayout'
-
-export const meta = {
-  id: 'introduction',
-  title: 'Supabase Handbook',
-  description: 'The Supabase public handbook and manifesto',
-}
-
-Supabase is thinking in public. Everything is online, no matter how raw the current state of things.
-
-This is our public handbook and manifesto. We are completely opensource and this means that you, our users, are also our team members.
-
----
-
-## Opensource
-
-Wherever we can we will support existing opensource tools and libraries.
-
-export const Page = ({ children }) => <Layout meta={meta} children={children} />
-
-export default Page
