+++ apps/docs/pages/reference/javascript/crawlers/[...slug].tsx
@@ -30,6 +30,7 @@ export default function JSReference(props) {
         typeSpec={typeSpec}
         pageProps={props}
         type="client-lib"
+        menuId="reference_javascript_v2"
       />
     </>
   )
