+++ apps/docs/pages/reference/kotlin/v1/[...slug].tsx
@@ -9,21 +9,22 @@ const sections = flattenSections(clientLibsCommonSections)
 const libraryPath = '/kotlin/v1'
 
 export default function KotlinReference(props) {
-  return (
-    <RefSectionHandler
-      sections={sections}
-      spec={spec}
-      pageProps={props}
-      type="client-lib"
-      isOldVersion
-    />
-  )
+	return (
+		<RefSectionHandler
+			menuId="reference_kotlin_v1"
+			sections={sections}
+			spec={spec}
+			pageProps={props}
+			type="client-lib"
+			isOldVersion
+		/>
+	)
 }
 
 export async function getStaticProps() {
-  return handleRefStaticProps(sections, libraryPath)
+	return handleRefStaticProps(sections, libraryPath)
 }
 
 export async function getStaticPaths() {
-  return handleRefGetStaticPaths(sections)
+	return handleRefGetStaticPaths(sections)
 }
