+++ packages/common/Providers.tsx
@@ -3,7 +3,7 @@ import { type ThemeProviderProps } from 'next-themes/dist/types'
 import { useEffect, useState } from 'react'
 
 export function ThemeProvider({ children, ...props }: ThemeProviderProps) {
-  const [mounted, setMounted] = useState(false)
+  const [mounted, setMounted] = useState(true) // temp, revert
   // @ts-ignore next-themes is old :/
 
   // useEffect only runs on the client, so now we can safely show the UI
